+++ packages/article-skeleton/src/article-body/drop-cap.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { fonts } from "@times-components/ts-styleguide";
+import { fontsWithFallback } from "@times-components/ts-styleguide";
 import { FontStorage, BoxExclusion } from "@times-components/typeset";
 import { TcText } from "@times-components/utils";
 import { getStringBounds } from "../body-utils";
@@ -16,7 +16,7 @@ export default (scale, color, dropCapFont, paragraph) => {
   const baseStyle = letter.attributes[0][0].settings;
   const fontSize = baseStyle.fontSize * 6;
   const fontSettings = {
-    fontFamily: fonts[dropCapFont],
+    fontFamily: fontsWithFallback[dropCapFont],
     fontStyle: "",
     fontWeight: "",
     fontSize,
@@ -44,7 +44,7 @@ export default (scale, color, dropCapFont, paragraph) => {
           left: styles.articleMainContentRow.paddingLeft,
           fontSize,
           lineHeight: height * 1.33,
-          fontFamily: fonts[dropCapFont],
+          fontFamily: fontsWithFallback[dropCapFont],
           color: fontSettings.color
         }
       ]}