+++ packages/pull-quote/src/styles/responsive.js
@@ -3,16 +3,16 @@ import styled from "styled-components";
 import {
   breakpoints,
   colours,
-  fonts,
+  fontsWithFallback,
   fontSizes
 } from "@times-components/ts-styleguide";
 
 const PullQuoteContent = styled(TcText)`
   color: ${colours.functional.primary};
-  font-family: "${fonts.headlineRegular}";
+  font-family: ${fontsWithFallback.headlineRegular};
   font-size: ${fontSizes.pageComponentHeadline}px;
   line-height: 30px;
-  
+
   cite {
     font-style: normal;
   }